Change sharpe_length type hint from float to int.

+++ src/pybroker/config.py
@@ -84,4 +84,4 @@ class StrategyConfig:
     exit_sell_fill_price: Union[
         PriceType, Callable[[str, BarData], Union[int, float, Decimal]]
     ] = field(default=PriceType.MIDDLE)
-    sharpe_length: Optional[float] = field(default=None)
+    sharpe_length: Optional[int] = field(default=None)
